+++ NEWS.md
@@ -1,4 +1,5 @@
 # QCkit 0.1.0.3
+28 February 2023 `te_check()` bug fix - exact column name filtering allows for multiple columns with similar names in the input data column. Improved documentation for transparency.
 23 February 2023
 * updated `te_check()`. It now supports searching multiple park units.
 22 February 2023